requirements.txt
alembic==1.10.4
loguru==0.7.0
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.11
typing_extensions==4.5.0
